begin vis refactor
